Move common files to root level
.github/: The GH CI workflows are merged together in a single yaml. More
change include the use of `working-directory` to make the job execute
in a subfolder. NOTE: GH currently does not support setting a workdir
globally for every step of a job. Other changes include using
actions/setup-node instead of borales/actions-yarn.

.gitignore: Most the .gitignore comes from backend/ with a few additions
from frontend/

.editorconfig: Moved from labextension/ to root, adding 4 spaces
indentation directive for python files.

docs/: This folders contained just the images used in the READMEs. The
kale banner is updated with the new logo and the labextension image is
updated with the new UI

README.md: The new README uses some sections from the previous ones but
most of it has been rewritten to reflect the current state of the
project.

docker/: The dev Dockerfiles are changed to install from a single
repository instead of two.

backend/Architecture.md and labextension/dev-notes.md are removed
because they have old and inaccurate information. They might be
reintroduced in the future with an updated documentation.

Common files that are just moved to the root of the project:

- LICENSE
- AUTHORS

Since backend/setup.py used backend/README.md to create a long
description for the PyPI package, this commit removes the long
descripton. This is not really needed for the moment so it does not
make sense to create an ad-hoc backend/README.md file.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>